# backend/src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FoodFlow API",
    "version": "1.0.0",
    "description": "API para o sistema FoodFlow, um SaaS para gestão de restaurantes. Inclui autenticação, gestão de restaurantes, usuários, mesas, cardápio, sessões, pedidos, relatórios, reservas e personalização de estilo."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    },
    {
      "url": "https://foodflow-api.render.com",
      "description": "Servidor de produção (Render)"
    }
  ],
...
  "paths": {
...
    "/tables": {
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Cria uma nova mesa",
        "description": "Cria uma nova mesa para um restaurante. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableInput"
              },
              "example": {
                "restaurant_id": "rest_1",
                "table_number": 1,
                "capacity": 4
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mesa criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                },
                "example": {
                  "id": "table_1",
                  "restaurant_id": "rest_1",
                  "table_number": 1,
                  "status": "available"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao criar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Mesa já existe"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{restaurantId}": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "Lista mesas de um restaurante",
        "description": "Retorna todas as mesas de um restaurante. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID do restaurante"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mesas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableResponse"
                  }
                },
                "example": [
                  {
                    "id": "table_1",
                    "restaurant_id": "rest_1",
                    "table_number": 1,
                    "status": "available"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Erro ao listar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Restaurante não encontrado"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{id}": {
      "put": {
        "tags": [
          "Tables"
        ],
        "summary": "Atualiza uma mesa",
        "description": "Atualiza o número ou status de uma mesa. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID da mesa"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTableInput"
              },
              "example": {
                "table_number": 2,
                "status": "occupied"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mesa atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                },
                "example": {
                  "id": "table_1",
                  "restaurant_id": "rest_1",
                  "table_number": 2,
                  "status": "occupied"
                }
              }
            }
          },
          "400": {
            "description": "Erro ao atualizar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Mesa não encontrada"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tables"
        ],
        "summary": "Exclui uma mesa",
        "description": "Remove uma mesa do sistema. Requer JWT.",
        "security": [
          {
            "bearerauth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID da mesa"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Mesa excluída"
          },
          "400": {
            "description": "Erro ao excluir",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Mesa não encontrada"
                }
              }
            }
          }
        }
      }
    },
    "/tables/validate-qr": {
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Valida um QR code",
        "description": "Valida o QR code de uma mesa, retornando informações da mesa para acesso ao cardápio.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "qr_code": {
                    "type": "string",
                    "description": "Código do QR code da mesa"
                  }
                },
                "required": [
                  "qr_code"
                ]
              },
              "example": {
                "qr_code": "table_1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "QR code validado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableResponse"
                },
                "example": {
                  "id": "table_1",
                  "restaurant_id": "rest_1",
                  "table_number": 1,
                  "status": "available"
                }
              }
            }
          },
          "400": {
            "description": "QR code inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "QR code inválido"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerauth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via /auth/login"
      }
    },
    "schemas": {
...
      "TableInput": {
        "type": "object",
        "properties": {
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante"
          },
          "table_number": {
            "type": "integer",
            "description": "Número da mesa"
          }
        },
        "required": [
          "restaurant_id",
          "table_number"
        ]
      },
      "TableResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID da mesa"
          },
          "restaurant_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID do restaurante"
          },
          "table_number": {
            "type": "integer",
            "description": "Número da mesa"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "reserved"
            ],
            "description": "Status da mesa"
          }
        },
        "required": [
          "id",
          "restaurant_id",
          "table_number",
          "status"
        ]
      },
      "UpdateTableInput": {
        "type": "object",
        "properties": {
          "table_number": {
            "type": "integer",
            "description": "Número da mesa",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "occupied",
              "reserved"
            ],
            "description": "Status da mesa",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Mensagem de erro específica"
                },
                "path": {
                  "type": "string",
                  "description": "Campo com erro"
                },
                "location": {
                  "type": "string",
                  "description": "Local do erro (ex.: body, query)"
                }
              }
            },
            "description": "Lista de erros detalhados",
            "nullable": true
          }
        }
      }
    }
  }
}